+++ include/openpose/gui/gui.hpp
@@ -6,7 +6,7 @@
 #include <openpose/core/common.hpp>
 #include <openpose/core/renderer.hpp>
 #include <openpose/gui/frameDisplayer.hpp>
-#include <openpose/pose/poseExtractor.hpp>
+#include <openpose/pose/poseExtractorNet.hpp>
 
 namespace op
 {
@@ -16,7 +16,7 @@ namespace op
         Gui(const Point<int>& outputSize, const bool fullScreen,
             const std::shared_ptr<std::atomic<bool>>& isRunningSharedPtr,
             const std::shared_ptr<std::pair<std::atomic<bool>, std::atomic<int>>>& videoSeekSharedPtr = nullptr,
-            const std::vector<std::shared_ptr<PoseExtractor>>& poseExtractors = {},
+            const std::vector<std::shared_ptr<PoseExtractorNet>>& poseExtractorNets = {},
             const std::vector<std::shared_ptr<Renderer>>& renderers = {});
 
         virtual void initializationOnThread();
@@ -34,7 +34,7 @@ namespace op
         // Frames display
         FrameDisplayer mFrameDisplayer;
         // Other variables
-        std::vector<std::shared_ptr<PoseExtractor>> mPoseExtractors;
+        std::vector<std::shared_ptr<PoseExtractorNet>> mPoseExtractorNets;
         std::vector<std::shared_ptr<Renderer>> mRenderers;
         std::shared_ptr<std::pair<std::atomic<bool>, std::atomic<int>>> spVideoSeek;
     };