+++ include/openpose/hand/headers.hpp
@@ -4,8 +4,8 @@
 // hand module
 #include <openpose/hand/handDetector.hpp>
 #include <openpose/hand/handDetectorFromTxt.hpp>
-#include <openpose/hand/handExtractor.hpp>
 #include <openpose/hand/handExtractorCaffe.hpp>
+#include <openpose/hand/handExtractorNet.hpp>
 #include <openpose/hand/handParameters.hpp>
 #include <openpose/hand/handCpuRenderer.hpp>
 #include <openpose/hand/handGpuRenderer.hpp>
@@ -15,7 +15,7 @@
 #include <openpose/hand/wHandDetectorFromTxt.hpp>
 #include <openpose/hand/wHandDetectorTracking.hpp>
 #include <openpose/hand/wHandDetectorUpdate.hpp>
-#include <openpose/hand/wHandExtractor.hpp>
+#include <openpose/hand/wHandExtractorNet.hpp>
 #include <openpose/hand/wHandRenderer.hpp>
 
 #endif // OPENPOSE_HAND_HEADERS_HPP