+++ src/openpose/gui/gui3D.cpp
@@ -417,10 +417,10 @@ namespace op
     Gui3D::Gui3D(const Point<int>& outputSize, const bool fullScreen,
                  const std::shared_ptr<std::atomic<bool>>& isRunningSharedPtr,
                  const std::shared_ptr<std::pair<std::atomic<bool>, std::atomic<int>>>& videoSeekSharedPtr,
-                 const std::vector<std::shared_ptr<PoseExtractor>>& poseExtractors,
+                 const std::vector<std::shared_ptr<PoseExtractorNet>>& poseExtractorNets,
                  const std::vector<std::shared_ptr<Renderer>>& renderers, const PoseModel poseModel,
                  const DisplayMode displayMode) :
-        Gui{outputSize, fullScreen, isRunningSharedPtr, videoSeekSharedPtr, poseExtractors, renderers},
+        Gui{outputSize, fullScreen, isRunningSharedPtr, videoSeekSharedPtr, poseExtractorNets, renderers},
         mDisplayMode{displayMode}
     {
         try